+++ servicecatalog_factory/tasks_unit_tests_helper.py
@@ -1,3 +1,6 @@
+# Copyright 2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# SPDX-License-Identifier: Apache-2.0
+
 import unittest
 from unittest import mock
