+++ servicecatalog_factory/template_builder/cdk/product_template.py
@@ -51,11 +51,15 @@ def create_cdk_pipeline(
         )
     )
     template.add_parameter(
-        t.Parameter("CDKSupportStartCDKDeployFunctionArn", Type="String",)
+        t.Parameter(
+            "CDKSupportStartCDKDeployFunctionArn",
+            Type="String",
+        )
     )
     template.add_parameter(
         t.Parameter(
-            "CDKSupportGetOutputsForGivenCodebuildIdFunctionArn", Type="String",
+            "CDKSupportGetOutputsForGivenCodebuildIdFunctionArn",
+            Type="String",
         )
     )
     template.add_parameter(
@@ -110,7 +114,9 @@ def create_cdk_pipeline(
         resource_type = "Custom::DeployDetails"
         props = dict()
 
-    runtime_versions = dict(nodejs=constants.BUILDSPEC_RUNTIME_VERSIONS_NODEJS_DEFAULT,)
+    runtime_versions = dict(
+        nodejs=constants.BUILDSPEC_RUNTIME_VERSIONS_NODEJS_DEFAULT,
+    )
     if configuration.get("runtime-versions"):
         runtime_versions.update(configuration.get("runtime-versions"))
 
@@ -124,7 +130,9 @@ def create_cdk_pipeline(
             ServiceRole=t.Sub(
                 "arn:aws:iam::${AWS::AccountId}:role${CDKSupportIAMRolePaths}${CDKSupportCDKDeployRoleName}"
             ),
-            Artifacts=codebuild.Artifacts(Type="NO_ARTIFACTS",),
+            Artifacts=codebuild.Artifacts(
+                Type="NO_ARTIFACTS",
+            ),
             Environment=codebuild.Environment(
                 ComputeType=t.Ref("CDKSupportCDKComputeType"),
                 EnvironmentVariables=[
@@ -193,7 +201,10 @@ def create_cdk_pipeline(
                                     ]
                                 },
                             ),
-                            artifacts={"name": "CDKDeploy", "files": ["*", "**/*"],},
+                            artifacts={
+                                "name": "CDKDeploy",
+                                "files": ["*", "**/*"],
+                            },
                         )
                     )
                 ),
@@ -210,7 +221,9 @@ def create_cdk_pipeline(
             ServiceRole=t.Sub(
                 "arn:aws:iam::${AWS::AccountId}:role${CDKSupportIAMRolePaths}${CDKSupportCDKDeployRoleName}"
             ),
-            Artifacts=codebuild.Artifacts(Type="NO_ARTIFACTS",),
+            Artifacts=codebuild.Artifacts(
+                Type="NO_ARTIFACTS",
+            ),
             Environment=codebuild.Environment(
                 ComputeType=t.Ref("CDKSupportCDKComputeType"),
                 EnvironmentVariables=[
@@ -278,7 +291,10 @@ def create_cdk_pipeline(
                                     ]
                                 },
                             ),
-                            artifacts={"name": "CDKDeploy", "files": ["*", "**/*"],},
+                            artifacts={
+                                "name": "CDKDeploy",
+                                "files": ["*", "**/*"],
+                            },
                         )
                     )
                 ),
@@ -307,7 +323,9 @@ def create_cdk_pipeline(
     template.add_resource(
         DeployDetailsCustomResource(
             "GetOutputsCode",
-            DependsOn=["StartCDKDeploy",],
+            DependsOn=[
+                "StartCDKDeploy",
+            ],
             ServiceToken=t.Ref("CDKSupportGetOutputsForGivenCodebuildIdFunctionArn"),
             CodeBuildBuildId=t.GetAtt("StartCDKDeploy", "BuildId"),
             BucketName=t.Sub("sc-cdk-artifacts-${AWS::AccountId}"),
